refactor(1014): extract APICoreFetchUtils to src/api/api_core_fetch_utils.py (P10, Cat E)
Position 10 of the 1014 hot-class refactor. Move APICoreFetchUtils to its own
module and keep a top-level re-export in MistHelper.py so callers keep working
without any callsite churn.

- New: src/api/api_core_fetch_utils.py (canonical body, 3 static methods)
- MistHelper.py: delete class body (was ~47 LOC), add re-export alias
- Lazy MistHelper resolution via importlib inside each method body
  (apisession + DEFAULT_API_PAGE_LIMIT) -- FR-028 IG-health clean
- pyproject.toml: add file to coverage.run omit list

Local gate: ruff clean, black clean.
FR-028 probe: MistHelper NOT in sys.modules after import of extracted module.
